+++ src/Entity/Simulator.php
@@ -20,6 +20,7 @@ class Simulator
      * @Assert\Length(max = 80, maxMessage = "Ce champs ne peux contenir plus de 80 caractères.");
      */
     private $firstName;
+
     /**
      * @var string
      * @Assert\NotBlank
@@ -28,12 +29,14 @@ class Simulator
      * @Assert\Length(max = 80,  maxMessage = "Ce champs ne peux contenir plus de 80 caractères.")
      */
     private $lastName;
+
     /**
      * @var string
      * @Assert\NotBlank
      * @Assert\Type("string")
      */
     private $address;
+
     /**
      * @var string
      * @Assert\NotBlank
@@ -42,34 +45,40 @@ class Simulator
      * @Assert\Length(max = 5, maxMessage = "Ce champs ne peux contenir plus de 5 chiffres.")
      */
     private $zipCode;
+
     /**
      * @var string
      * @Assert\NotBlank
      * @Assert\Type("string")
      */
     private $city;
+
     /**
      * @var string
      * @Assert\NotBlank
      */
     private $zone;
+
     /**
      * @var \DateTime
      * @Assert\NotBlank
      * @Assert\Type("\DateTime")
      */
     private $acquisitionDate;
+
     /**
      * @var int
      * @Assert\Type("int")
      */
     private $duration;
+
     /**
      * @var float
      * @Assert\NotBlank
      * @Assert\Type("float")
      */
     private $surfaceArea;
+
     /**
      * @var int
      * @Assert\NotBlank
@@ -77,6 +86,7 @@ class Simulator
      *
      */
     private $purchasePrice;
+
     /**
      * @var int
      * @Assert\NotBlank
@@ -84,6 +94,7 @@ class Simulator
      * @Assert\GreaterThan(0)
      */
     private $notaryFees;
+
     /**
      * @var int
      * @Assert\NotBlank
@@ -91,6 +102,7 @@ class Simulator
      * @Assert\GreaterThan(0)
      */
     private $otherFeesAcquisition;
+
     /**
      * @var int
      * @Assert\NotBlank
@@ -228,17 +240,17 @@ class Simulator
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function getSurfaceArea(): ?int
+    public function getSurfaceArea(): ?float
     {
         return $this->surfaceArea;
     }
 
     /**
-     * @param int $surfaceArea
+     * @param float $surfaceArea
      */
-    public function setSurfaceArea(int $surfaceArea): void
+    public function setSurfaceArea(float $surfaceArea): void
     {
         $this->surfaceArea = $surfaceArea;
     }
